match flags with existing kubectl + helm flags (#26)
The `--chart-dir` flag is now also available as `-f` to better integrate
with the `kubectl apply -f` workflow. Also the `--value-file` flag was
renamed to `--values` to match the flag helm uses for the same purpose.
To avoid confusion, the `--hook` flag was renamed to `--hook-type` to
make clean that a hook type is expected and not a hook name.

This change also updates the roadmap in README.md with some planned
features.